requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
requests==2.26.0
pydicom==2.2.2
Pillow==8.3.2
boto3==1.18.50
python-multipart==0.0.5